MAINT: shuffled code around
